Move ndcube slice and crop tests to own module.
